First working version

+++ inc/labspectreipc.h
@@ -48,10 +48,10 @@ typedef struct spectre_lab_command_t {
 	spectre_lab_command_kind kind;
 
 	// Usually this is the user virtual address of the shared memory region
-	uint32_t arg1;
+	uint64_t arg1;
 
 	// Usually this is the offset into the secret to access
-	uint32_t arg2;
+	uint64_t arg2;
 } spectre_lab_command;
 
 #endif // SHD_SPECTRE_LAB_IPC_H
